third_party/blink/web_tests/external/wpt/upgrade-insecure-requests/gen/top.http-rp/upgrade/sharedworker-module.https-expected.txt

This is a testharness.js-based test.
FAIL Upgrade-Insecure-Requests: Expects allowed for sharedworker-module to same-http-downgrade origin and downgrade redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
FAIL Upgrade-Insecure-Requests: Expects allowed for sharedworker-module to same-http-downgrade origin and no-redirect redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
PASS Upgrade-Insecure-Requests: Expects allowed for sharedworker-module to same-https origin and downgrade redirection from https context.
Harness: the test ran to completion.
